Merge branch "dev" into branch "ui"

+++ tests/core/reversescan/main.py
@@ -1,3 +1,4 @@
+from re import search
 from time import sleep
 from fastapi import FastAPI
 from os import getenv
@@ -7,15 +8,19 @@ from traceback import format_exc
 from uvicorn import run
 
 
-app = FastAPI()
-fastapi_proc = Process(target=run, args=(app,), kwargs=dict(host="0.0.0.0", port=80))
-fastapi_proc.start()
+fastapi_proc = None
+if getenv("TEST_TYPE", "docker") == "docker":
+    app = FastAPI()
+    fastapi_proc = Process(
+        target=run, args=(app,), kwargs=dict(host="0.0.0.0", port=80)
+    )
+    fastapi_proc.start()
 
-sleep(1)
+    sleep(1)
 
 try:
-    use_reverse_scan = getenv("USE_REVERSE_SCAN", "no") == "yes"
-    reverse_scan_ports = getenv("REVERSE_SCAN_PORTS", "22 80 443 3128 8000 8080")
+    use_reverse_scan = getenv("USE_REVERSE_SCAN", "yes") == "yes"
+    reverse_scan_ports = getenv("REVERSE_SCAN_PORTS", "80")
 
     print(f"ℹ️ Trying to access http://www.example.com ...", flush=True)
     status_code = get(
@@ -26,7 +31,7 @@ try:
 
     if status_code == 403:
         pass
-    elif use_reverse_scan and " 80 " in reverse_scan_ports:
+    elif use_reverse_scan and search(r"\b80\b", reverse_scan_ports):
         print(
             "❌ Request didn't return 403, but reverse scan is enabled and port 80 is in the reverse scan ports list, exiting ...",
             flush=True,
@@ -40,4 +45,5 @@ except:
     print(f"❌ Something went wrong, exiting ...\n{format_exc()}", flush=True)
     exit(1)
 finally:
-    fastapi_proc.terminate()
+    if fastapi_proc:
+        fastapi_proc.terminate()
